refactor: move bundler files into own directory
