+++ api/src/bundler/getPlugins.ts
@@ -1,9 +1,9 @@
-import { theme } from '../utils/plugins/codeHikeTheme.js';
+import { theme } from './plugins/codeHikeTheme.js';
 import { remarkCodeHike } from '@code-hike/mdx';
 import remarkGfm from 'remark-gfm';
-import rehypeCodeBlocks from '../utils/plugins/rehype-code-blocks.js';
+import rehypeCodeBlocks from './plugins/rehype-code-blocks.js';
 import { rehypeAccessibleEmojis } from 'rehype-accessible-emojis';
-import rehypeInlineBadges from '../utils/plugins/rehype-inline-badges.js';
+import rehypeInlineBadges from './plugins/rehype-inline-badges.js';
 import rehypeSlug from 'rehype-slug';
 import remarkMath from 'remark-math';
 import rehypeKatex from 'rehype-katex';
