+++ api/src/bundler/plugins/remark-component-check.ts
