+++ api/src/bundler/plugins/remark-undeclared-variables.ts
